+++ travis/push-doc-to-gh-pages.sh
@@ -5,7 +5,7 @@ echo -e "TRAVIS_PULL_REQUEST: $TRAVIS_PULL_REQUEST\n"
 echo -e "TRAVIS_BRANCH: $TRAVIS_BRANCH\n"
 echo -e "TRAVIS_TAG: $TRAVIS_TAG\n"
 
-if [ "$TRAVIS_PULL_REQUEST" == "false" ] && [[ "$TRAVIS_BRANCH" =~ ^(master|.*0\.x)$ ]] && [ "$TRAVIS_TAG" == "" ]; then
+if [ "$TRAVIS_PULL_REQUEST" == "false" ] && [[ "$TRAVIS_BRANCH" =~ ^(master|.*\.x)$ ]] && [ "$TRAVIS_TAG" == "" ]; then
   case "${TRAVIS_JOB_NUMBER}" in
        *\.1)
         echo -e "Invoking auto-doc deployment for Travis job ${TRAVIS_JOB_NUMBER}"